% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getuniquesolution.R
\name{getuniquesolution}
\alias{getuniquesolution}
\title{getuniquesolution}
\usage{
getuniquesolution(Values)
}
\arguments{
\item{Values}{The output from getvalue}
}
\value{
a vector. which is the solution of the fixed points.
}
\description{
function that solve the equation syseq(m)=m, and return m*1i.
}
\examples{
f1<-function(x){x};  Values=getvalue(0.001,10/3,1,0.1,f1,psi1=1); getuniquesolution(Values)
}
