% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getvalue.R
\name{getvalue}
\alias{getvalue}
\title{getvalue}
\usage{
getvalue(lbd, psi, F1, tau, ...)
}
\arguments{
\item{lbd}{a real number >0, the tuning parameter of ridge regression.}

\item{psi}{n/d, n is sample size and d is sample dimension.}

\item{F1}{larger or equal to 0, the norm of beta.}

\item{tau}{standard deviation of noise.}

\item{...}{Input all activation functions and N_j/d.}
}
\value{
a list, contain all hyper parameters.
}
\description{
a function that ensemble all hyper parameters in MRFM.
}
\details{
It is a function that ensemble all hyper parameters in MRFM. Users may use this function to give a precise MRFM model.
}
\examples{
f1<-function(x){x};  Values=getvalue(0.001,10/3,1,0.1,f1,psi1=1)
f1<-function(x){x};  Values=getvalue(0.001,10/3,1,0.1,f1,psi1=1,ReLU,psi2=2)
}
