% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainvalue.R
\name{mainvalue}
\alias{mainvalue}
\title{mainvalue}
\usage{
mainvalue(Values)
}
\arguments{
\item{Values}{The output from getvalue.}
}
\value{
a number: theoretical value of the specific MRFM.
}
\description{
a function that ensemble all hyper parameters in MRFM.
}
\details{
It is a function that ensemble all hyper parameters in MRFM. Users may use this function to give a precise MRFM model.
}
\examples{
f1<-function(x){x};  Values=getvalue(0.001,10/3,1,0.1,f1,psi1=1);mainvalue(Values)
f1<-function(x){x};  Values=getvalue(0.001,10/3,1,0.1,f1,ReLU,psi1=1,psi2=1.5);mainvalue(Values)
}
