% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixV.R
\name{MatrixV}
\alias{MatrixV}
\title{MatrixV}
\usage{
MatrixV(nu, Values)
}
\arguments{
\item{nu}{the fixed points obtained from getuniquesolution}

\item{Values}{The output from getvalue.}
}
\value{
a matrix V.
}
\description{
a function that return matrix V in MRFM.
}
